# orchestrator/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Steel Browser Instance API",
    "description": "Documentation for controlling a single instance of Steel Browser",
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
...
    "/sessions/{sessionId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/sessions/{sessionId}/status": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sessionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ]
}